migrated ELR fn out
